+++ internal/provider/provider.go
@@ -374,6 +374,7 @@ func (p *KonnectProvider) Resources(ctx context.Context) []func() resource.Resou
 
 func (p *KonnectProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		NewCloudGatewayNetworkDataSource,
 		NewCloudGatewayProviderAccountListDataSource,
 		NewGatewayControlPlaneDataSource,
 		NewGatewayControlPlaneListDataSource,